Add `memory64` crate feature to Wasmi (#1934)
* add new memory64 crate feature to wasmi

* optimize effective_address for disabled memory64

* add wasmi_core/memory64 crate feature

* forbig 64-bit memories when memory64 is disabled

* cfg-guard Config::wasm_memory64 + update default

* cfg-guard MemoryType::new64 constructor

* cfg-ignore test that requires memory64

* add and forward memory64 feature to wasmi_cli

* enable memory64 feature in wasmi_fuzz

* enable memory64 in wasmi_wast

* enable memory64 in wasmi fuzzing

* add memory64 to wasmi_c_api default feature set

* add new Offset type to wasmi_ir

* add memory64 crate feature to wasmi_ir

* remove memory64 comment. put docs elsewhere.

* use Offset type for offset field of load/store ops

* update op_size_of_and_alignment test

* add GetValue support for Offset

* use Offset in load/store ops definitions of the translator

* adjust simd translation logic

* support new Offset in translator

* replace constructors with {Offset16,Address}::new

* update translator to reflect constructor changes

This allows putting the Offset::new check to the front of the translation logic.

* refactor decode_memarg to return new Offset type

* simplify Offset[16] APIs

* move memory64 feature docs to crate docs

* remove memory64 from base feature set of wasmi_cli

* add comment about memory64 for embedded users

+++ crates/wasmi/src/engine/executor/handler/exec/simd.rs
@@ -556,7 +556,7 @@ macro_rules! handler_store_lane_mem0_offset16_ss {
                     let offset = get_value(offset, sp, ireg, freg32, freg64);
                     let value = get_value(value, sp, ireg, freg32, freg64);
                     let mem_bytes = $crate::engine::executor::handler::state::mem0_bytes(mem0, mem0_len);
-                    $eval(mem_bytes, ptr, u64::from(u16::from(offset)), value, lane).into_control()?;
+                    $eval(mem_bytes, ptr, u64::from(offset), value, lane).into_control()?;
                     dispatch!(state, ip, sp, mem0, mem0_len, instance, ireg, freg32, freg64)
                 }
             }
